Init project and restructure
