+++ NimbleExtension/Supporting Files/NimbleExtension.h
@@ -1,0 +1,19 @@
+//
+//  NimbleExtension.h
+//  NimbleExtension
+//
+//  Created by Su Van Ho on 21/8/19.
+//  Copyright © 2019 Nimble. All rights reserved.
+//
+
+#import <UIKit/UIKit.h>
+
+//! Project version number for NimbleExtension.
+FOUNDATION_EXPORT double NimbleExtensionVersionNumber;
+
+//! Project version string for NimbleExtension.
+FOUNDATION_EXPORT const unsigned char NimbleExtensionVersionString[];
+
+// In this header, you should import all the public headers of your framework using statements like #import <NimbleExtension/PublicHeader.h>
+
+
